DRY up linode views
