+++ src/features/linodes/LinodesLanding/LinodeCard.tsx
@@ -1,4 +1,4 @@
-import { compose, splitAt } from 'ramda';
+import { compose } from 'ramda';
 import * as React from 'react';
 import { Link } from 'react-router-dom';
 
@@ -16,8 +16,6 @@ import Typography from '@material-ui/core/Typography';
 import Flag from 'src/assets/icons/flag.svg';
 import CircleProgress from 'src/components/CircleProgress';
 import Grid from 'src/components/Grid';
-import ShowMore from 'src/components/ShowMore';
-import Tag from 'src/components/Tag';
 import { withTypes } from 'src/context/types';
 import { LinodeConfigSelectionDrawerCallback } from 'src/features/LinodeConfigSelectionDrawer';
 import { linodeInTransition, transitionText } from 'src/features/linodes/transitions';
@@ -216,6 +214,7 @@ interface Props {
   openConfigDrawer: (configs: Linode.Config[], action: LinodeConfigSelectionDrawerCallback) => void;
   toggleConfirmation: (bootOption: Linode.BootAction,
     linodeId: number, linodeLabel: string) => void;
+  renderTagsAndMoreTags: (tags: string[]) => JSX.Element;
 }
 
 interface TypesContextProps {
@@ -227,8 +226,6 @@ type CombinedProps =
   Props &
   TypesContextProps &
   WithStyles<CSSClasses>;
-
-const tagsToShow = 3;
 
 class LinodeCard extends React.Component<CombinedProps, State> {
   state: State = {
@@ -315,8 +312,6 @@ class LinodeCard extends React.Component<CombinedProps, State> {
       typesData,
     } = this.props;
 
-    const [visibleTags, additionalTags] = splitAt(tagsToShow, linodeTags);
-
     return (
       <CardContent className={`${classes.cardContent} ${classes.customeMQ}`}>
         <div>
@@ -335,42 +330,11 @@ class LinodeCard extends React.Component<CombinedProps, State> {
             {imageLabel}
           </div>
           <div className={classes.cardSection}>
-            {
-              visibleTags.map(eachTag => {
-                return (
-                  <Tag
-                    label={eachTag}
-                    key={eachTag}
-                    variant="gray"
-                  />
-                )
-              })
-            }
-            {additionalTags && this.renderMoreTags(additionalTags)}
+            {this.props.renderTagsAndMoreTags(linodeTags)}
           </div>
         </div>
       </CardContent>
     );
-  }
-
-  renderTag = (tags: string[]) => {
-    const { classes } = this.props;
-    return tags.map(eachTag => {
-      return (
-        <div key={eachTag} className={classes.row}>
-          <div className={`${classes.tag}`}>{eachTag}</div>
-        </div>
-      )
-    })
-  }
-
-  renderMoreTags = (tags: string[]) => {
-    return (
-      <ShowMore
-        items={tags}
-        render={this.renderTag}
-      />
-    )
   }
 
   renderFlag = () => {
